Seidel Space

We follow [Iri17, Definition 3.2] in our description of the Seidel space. Let $X$ be a (smooth) GKM variety with torus action by $T$. Let $\iota\colon\mathbb{C}^\times\rightarrow T$ be a group homomorphism.

The Seidel space $S_X$ associated to this datum is a $X$-bundle over $\mathbb{P}^1$ which is trivialized over

\[ \begin{align*} U_0 &= \{[1:u] : u\in\mathbb{C}\}\subset\mathbb{P}^1\\ U_\infty &= \{[v:1] : v\in\mathbb{C}\}\subset\mathbb{P}^1. \end{align*}\]

The transition function on $U_0\cap U_\infty$ is given by

\[ U_0\times X \ni ([1:u], x) \longmapsto ([u^{-1}:1], \iota(u)\cdot x) \in U_\infty\times X.\]

Globally, it can be described as

\[ S_X = \left( \mathbb{C}^2\setminus\{0\}\times X \right) / \mathbb{C}^\times,\]

where $\lambda\in\mathbb{C}^\times$ acts by $\lambda\cdot ((v, u), x) := ((\lambda v, \lambda u), \iota(\lambda)\cdot x)$.

Let $\widehat{T}:= T\times\mathbb{C}^\times$. Then $\widehat{T}$ acts on $S_X$ by $(t, s)\cdot ([(v, su), t\cdot x])$.

That is, $\widehat{T}$ acts on $S_X\rightarrow\mathbb{P}^1$ via the $T$-action on the fibres $X$ and via the standard $\mathbb{C}^\times$-action on the base $\mathbb{P}^1$.

Note

If $X$ is a GKM space with respect to $T$, then $S_X$ is a GKM space with respect to $\widehat{T}$, which is implemented in this package.

The GKM graph and connection of $S_X$ is described in the supporting paper.

GKMtools.Seidel_space — Function

Seidel_space(G::AbstractGKM_graph, w::AbstractAlgebra.Generic.FreeModuleElem{R}; basePoint::Int64 = 1) -> AbstractGKM_graph

Construct the Seidel space associated to the GKM graph G (representing the GKM variety $X$) and the map $\iota:\mathbb{C}^\times\rightarrow T$ given by the element $w\in\mathfrak{t}$.

Optional arguments:

basePoint::Int64: This is the vertex of G so that in the internal presentation of the curve classes of $S_X$, the curve class of the section of $S_X\rightarrow \mathbb{P}^1$ associated to the vertex basePoint is represented by (0,...,0,1). The first entries correspond to curve classes of $X$. The last is the degree of the curve class projected to $\mathbb{P}^1$.

Examples

julia> G = projective_space(GKM_graph, 2);

julia> S = Seidel_space(G, gens(G.M)[1])
GKM graph with 6 nodes, valency 3 and axial function:
[2]_0 -> [1]_0 => (-1, 1, 0, 0)
[3]_0 -> [1]_0 => (-1, 0, 1, 0)
[3]_0 -> [2]_0 => (0, -1, 1, 0)
[1]_inf -> [1]_0 => (0, 0, 0, -1)
[2]_inf -> [2]_0 => (0, 0, 0, -1)
[2]_inf -> [1]_inf => (-1, 1, 0, 1)
[3]_inf -> [3]_0 => (0, 0, 0, -1)
[3]_inf -> [1]_inf => (-1, 0, 1, 1)
[3]_inf -> [2]_inf => (0, -1, 1, 0)

julia> S = Seidel_space(G, gens(G.M)[1] + 7*gens(G.M)[2])
GKM graph with 6 nodes, valency 3 and axial function:
[2]_0 -> [1]_0 => (-1, 1, 0, 0)
[3]_0 -> [1]_0 => (-1, 0, 1, 0)
[3]_0 -> [2]_0 => (0, -1, 1, 0)
[1]_inf -> [1]_0 => (0, 0, 0, -1)
[2]_inf -> [2]_0 => (0, 0, 0, -1)
[2]_inf -> [1]_inf => (-1, 1, 0, -6)
[3]_inf -> [3]_0 => (0, 0, 0, -1)
[3]_inf -> [1]_inf => (-1, 0, 1, 1)
[3]_inf -> [2]_inf => (0, -1, 1, 7)

julia> print_curve_classes(S)
[2]_0 -> [1]_0: (1, 0), Chern number: 3
[3]_0 -> [1]_0: (1, 0), Chern number: 3
[3]_0 -> [2]_0: (1, 0), Chern number: 3
[1]_inf -> [1]_0: (0, 1), Chern number: -3
[2]_inf -> [2]_0: (6, 1), Chern number: 15
[2]_inf -> [1]_inf: (1, 0), Chern number: 3
[3]_inf -> [3]_0: (-1, 1), Chern number: -6
[3]_inf -> [1]_inf: (1, 0), Chern number: 3
[3]_inf -> [2]_inf: (1, 0), Chern number: 3

Using a different base point does not change the resulting GKM graph but gives a different internal presentation of the curve classes.

julia> S = Seidel_space(G, gens(G.M)[1] + 7*gens(G.M)[2]; basePoint=2)
GKM graph with 6 nodes, valency 3 and axial function:
[2]_0 -> [1]_0 => (-1, 1, 0, 0)
[3]_0 -> [1]_0 => (-1, 0, 1, 0)
[3]_0 -> [2]_0 => (0, -1, 1, 0)
[1]_inf -> [1]_0 => (0, 0, 0, -1)
[2]_inf -> [2]_0 => (0, 0, 0, -1)
[2]_inf -> [1]_inf => (-1, 1, 0, -6)
[3]_inf -> [3]_0 => (0, 0, 0, -1)
[3]_inf -> [1]_inf => (-1, 0, 1, 1)
[3]_inf -> [2]_inf => (0, -1, 1, 7)

julia> print_curve_classes(S)
[2]_0 -> [1]_0: (1, 0), Chern number: 3
[3]_0 -> [1]_0: (1, 0), Chern number: 3
[3]_0 -> [2]_0: (1, 0), Chern number: 3
[1]_inf -> [1]_0: (-6, 1), Chern number: -3
[2]_inf -> [2]_0: (0, 1), Chern number: 15
[2]_inf -> [1]_inf: (1, 0), Chern number: 3
[3]_inf -> [3]_0: (-7, 1), Chern number: -6
[3]_inf -> [1]_inf: (1, 0), Chern number: 3
[3]_inf -> [2]_inf: (1, 0), Chern number: 3