Seidel elements / Shift operators

This section deals with certain equivariant Gromov–Witten invariants on Seidel spaces. The definition used in this package should be carefully compared to [Iri17, Section 3] and [MO19, Chapter 8] (shift operators). A (non-equivariant) symplectic account can be found in [MS12, Section 11.4] (Seidel representation).

Let $X$ be a (smooth projective) GKM variety with torus action by $T$, and let $\iota\colon \mathbb{C}^\times \rightarrow T$ be a group homomorphism. This gives rise to the $X$-bundle $\pi\colon S_X\rightarrow \mathbb{P}^1$, where $S_X$ is the Seidel space associated to $(X, \iota)$. Recall that $S_X$ is a GKM space with respect to $\widehat{T}:= T\times\mathbb{C}^\times$, where the extra copy of $\mathbb{C}^\times$ acts by rotating the base $\mathbb{P}^1$.

Let $(e_i)$ and $(e^i)$ be dual bases of $H_T^*(X;\mathbb{Q})$ with respect to the $T$-equivariant Poincaré pairing on $X$. Note also that we have a $H_T^*(\text{pt};\mathbb{Q})$-linear pushforward map $(i_\infty)_*\colon H_T^*(X;\mathbb{Q}) \rightarrow H_{\widehat{T}}^*(S_X;\mathbb{Q})$ raising degree by one.

Finally, let $H_2^\text{sec}(S_X;\mathbb{Z})$ be the set of effective section curve classes in $S_X$, i.e., effective curve classes that project to $[\mathbb{P}^1]$ under $\pi\colon S_X\rightarrow\mathbb{P}^1$. The additive group $H_2^\text{sec}(S_X;\mathbb{Z})$ is an $H_2^\text{eff}(X;\mathbb{Z})$-torsor, so after picking some $\beta_0\in H_2^\text{sec}(S_X;\mathbb{Z})$ there is an identification $r\colon H_2^\text{sec}(S_X;\mathbb{Z})\stackrel{\cong}{\longrightarrow} H_2^\text{eff}(X;\mathbb{Z})$ that sends $\beta_0$ to $0$.

The (equivariant) Seidel element associated to $\iota$ is

\[ \mathcal{S}(\iota) := \sum_{\beta\in H_2^\text{sec}(S_X;\mathbb{Z})} \left[GW^{\widehat{T}}_{0,1,\beta}((i_\infty)_*(e_i))\right]_{\hat{t} = 0} e^i q^{r(\beta)} \in QH_T^*(X)\]

where $\hat{t}$ is the equivariant parameter for the extra $\mathbb{C}^\times$ in $\widehat{T}$.

Note

Let $\Lambda:= \text{Hom}(\mathbb{C}^\times, T)$ be the cocharacter lattice of $T$. It is a key property of the equivariant Seidel elements that the map

\[ \mathcal{S}\colon \Lambda \rightarrow QH_T^*(X)^\times\]

is a group homomorphism, where $QH_T^*(X)^\times$ is endowed with the equivariant quantum product.

GKMtools.Seidel_element — Function

Seidel_element(G::AbstractGKM_graph, w::AbstractAlgebra.Generic.FreeModuleElem{R}, cMax::Int64) where R<:GKM_weight_type

Return the Seidel element on G associated to the homomorphism $\iota\colon \mathbb{C}^\times\rightarrow T$ defined by w. This calculates the contribution of curve classes with Chern number at most cMax.

Warning

This requires is_strictly_NEF(G) == true as otherwise there might be infinitely many contributing curve classes.

Curve classes

The curve classes of the Seidel space are represented as $(\beta, n)$ where $\beta$ is a curve class of $X$ and the integer $n$ is the degree of the curve projected along $S_X\rightarrow\mathbb{P}^1$.

The inclusions of the fibre $X$ over $0$ and $\infty$ simply send $\beta \mapsto (\beta,0)$.

Each fixed point $v\in X^T$ gives an invariant curve that is a section of $S_X\rightarrow \mathbb{P}^1$. Its curve class is $(\beta_v,1)$, where the $\beta_v$ are only uniquely determined up to an overall constant. The Seidel space used in this function is normalized such that $\beta_1=0$, where $1$ is the fixed point represented by the GKM graph vertex with index 1.

Example

julia> P1 = projective_space(GKM_graph, 1);

julia> t1, t2 = gens(P1.M);

julia> cMax = 6;

julia> S1 = Seidel_element(P1, t1, cMax)
(0, -t1 + t2) q^(-1)

julia> S2 = Seidel_element(P1, t2, cMax)
(t1 - t2, 0) q^(0)

julia> S3 = Seidel_element(P1, t1+t2, cMax)
(1, 1) q^(0)

julia> S1 * S2 == S3
true

julia> S4 = Seidel_element(P1, -2*t1, cMax)
(t1^2 - 2*t1*t2 + t2^2, 0) q^(0)
 + (1, 1) q^(1)

julia> S4 * S1 * S1
(1, 1) q^(0)