Miscellaneous

In this section we collect functions that are useful for GKM graphs, but that do not belong to GKM theory.

Smoothness of Schubert varieties

GKMtools.isregular_word — Function

isregular_word(w::WeylGroupElem) -> Bool

Given a Weyl group element w, this function returns true if and only if the Bruhat graph of w is regular. That is, all vertices have the same number of incident edges. We implemented [BP05, Lemma 5.7]. Note that the Schubert variety given by w is rationally smooth if and only if the Bruhat graph of w is regular.

GKMtools.issmooth_schubert_at_v — Function

issmooth_schubert_at_v(v::WeylGroupElem, w::WeylGroupElem) -> Bool

Given a Weyl group element w, this function returns true if and only if the Schubert variety given by w is smooth at the point given by v. We implemented [Kum96, Theorem 5.5], see also [BL00, 7.2.1 Theorem].

GKMtools.issmooth_schubert — Function

issmooth_schubert(w::WeylGroupElem) -> Bool

Check if the Schubert variety given by w is smooth at the point given by the identity element. This is equivalent to the Schubert variety be smooth.

Kazhdan-Lusztig polynomials

GKMtools.R_polynomial — Function

R_polynomial(v::WeylGroupElem, w::WeylGroupElem) -> ZZPolyRingElem

Given two Weyl group elements v,w, with v≤w this function returns the R-polynomial of v,w in the polynomial ring $\mathbb{Z}[q]$. See [BL00, Section 6.1].

GKMtools.kazhdan_lusztig — Function

kazhdan_lusztig(v::WeylGroupElem, w::WeylGroupElem) -> ZZPolyRingElem

Given two Weyl group elements v,w, with v≤w this function returns the Kazhdan-Lusztig polynomial of v,w in the polynomial ring $\mathbb{Z}[q]$. See [BL00, Equation (6.1.8)].