Equivariant Gromov–Witten invariants of GKM graphs

One of the main features of this package is to calculate equivariant Gromov-Witten invariants of GKM spaces $X$ (see Definition).

Mathematical ingredients: The following mathematical ingredients are used in the package:

The localization formula from [LS17, Hir23] that expresses equivariant Gromov-Witten invariants in the form

\[ GW^T_{X,g=0,m,\beta}(...) = \sum_{\overrightarrow{\Gamma}} GW^T_{\overrightarrow{\Gamma}}\]

where the sum is over certain decorated trees $\overrightarrow{\Gamma}$ mapping into the GKM graph of $X$, such that the edge multiplicities sum to the curve class $\beta\in H_2(X;\mathbb{Z})$. Here, $GW^T_{\overrightarrow{\Gamma}}$ is a rational function in $\dim{T}$ many variables, i.e. an element of $\text{Frac}H_T^*(\text{pt};\mathbb{Q})$.

An efficient way of enumerating trees with marked points and counting their automorphisms (inspired by [WROM86]).

A way of enumerating all combination of edges with multiplicities in the GKM graph that sum to a given curve class $\beta\in H_2(X;\mathbb{Z})$ (see supporting paper).

Integrating on the moduli space

GKMtools.gromov_witten — Function

gromov_witten(G::AbstractGKM_graph, beta::CurveClass_type, n_marks::Int64, P_input::EquivariantClass; show_bar::Bool = false) -> GW invariants

Integrate the class P_input over the moduli space $\overline{\mathcal{M}_{0,n}}(X,\beta)$ of genus 0 stable maps to $X$ in class $\beta\in H_2(X;\mathbb{Z})$ with n_marks marked points. The result is an element of $\text{Frac}(H_T^*(\text{pt};\mathbb{Q}))$, i.e. a rational function in $\dim_\mathbb{C}(T)$ many variables.

Note

If the underlying space is a (smooth projective or Hamiltonian) GKM space then the output should in fact live in $H_T^*(\text{pt};\mathbb{Q})$, so it should be a polynomial in the $\dim_\mathbb{C}(T)$ many variables.

Warning

The GKM graph G must have a connection, as this datum is required by the localization formula [LS17].

Arguments

G::AbstractGKM_graph: The GKM graph of the target GKM space $X$.

beta::CurveClass_type: The (non-zero) curve class $\beta\in H_2(X;\mathbb{Z})$ in which the image of the stable map should lie. To produce beta, use functions like curve_class (see Curve Classes).

P_input::EquivariantClass: The equivariant cohomology class on $\overline{\mathcal{M}_{0,n}}(X,\beta)$ that is being integrated. Use the functions ev, class_one, and Psi to produce this. These classes also support arithmetic using +, *, et cetera.

show_bar::Bool: If true, a progress bar will be displayed showing the estimated time until completion. This should be used for big examples.

fast_mode::Bool: If the expected result of the computation is a number, this option will speed up the computation.

Warning

If the expected result of the computation is not a number and fast_mode is true, the result will be a meaningless number.

Example

julia> P2 = projective_space(GKM_graph, 2);

julia> beta = curve_class(P2, Edge(1, 2));

julia> gromov_witten(P2, beta, 1, ev(1, point_class(P2, 1)); show_bar=false)
0

julia> gromov_witten(P2, beta, 2, ev(1, point_class(P2, 1)) * ev(2, point_class(P2, 2)); show_bar=false)
1

julia> gromov_witten(P2, beta, 2, ev(1, point_class(P2, 1)) * ev(2, point_class(P2, 1)); show_bar=false, fast_mode=true)
1

julia> gromov_witten(P2, beta, 2, ev(1, point_class(P2, 1))^2 * ev(2, point_class(P2, 2)); show_bar=false)
t1^2 - t1*t2 - t1*t3 + t2*t3

julia> gromov_witten(P2, beta, 3, ev(1, point_class(P2, 1)) * ev(2, point_class(P2, 1)) * ev(3, point_class(P2, 3)); show_bar=false)
t1 - t2

gromov_witten(V::GKM_vector_bundle, beta::CurveClass_type, n_marks::Int64, P_input::EquivariantClass; show_bar::Bool = false, check_degrees::Bool = false)

Same as before, but taking the total space of a GKM vector bundle as input.

gromov_witten(V::GKM_vector_bundle, beta::CurveClass_type, n_marks::Int64, P_input::EquivariantClass; show_bar::Bool = false, check_degrees::Bool = false)

Same as before, but taking the total space of a GKM vector bundle as input.

Cohomology classes on the moduli space

GKMtools.ev — Function

ev(j::Int64, cc) -> EquivariantClass

Return the equivariant cohomology class on $\overline{\mathcal{M}_{0,n}}(X,\beta)$ that is given by pulling back the cohomology class cc on $X$ along the evaluation map at the j-th point: $\text{ev}_j\colon \overline{\mathcal{M}_{0,n}}(X,\beta)\rightarrow X$.

Example

julia> G24 = grassmannian(GKM_graph, 2, 4);

julia> e1 = ev(1, point_class(G24, 1));

julia> e2 = ev(2, first_chern_class(G24));

julia> e3 = ev(3, poincare_dual(gkm_subgraph_from_vertices(G24, [1, 2])));

julia> beta = curve_class(G24, Edge(1, 2));

julia> gromov_witten(G24, beta, 3, e1 * e2 * e3; show_bar=false)
4

GKMtools.class_one — Function

class_one() -> EquivariantClass

Return the cohomology class $1$ on $\overline{\mathcal{M}_{0,n}}(X,\beta)$.

Example

julia> P2 = projective_space(GKM_graph, 2);

julia> beta = curve_class(P2, Edge(1, 2));

julia> gromov_witten(P2, beta, 2, ev(1, point_class(P2, 1)) * ev(2, point_class(P2, 2)); show_bar=false)
1

julia> gromov_witten(P2, beta, 2, ev(1, point_class(P2, 1)) * ev(2, point_class(P2, 2)) * class_one(); show_bar=false)
1

julia> gromov_witten(P2, beta, 0, class_one(); show_bar=false)
0

GKMtools.Psi — Function

Psi(a) -> EquivariantClass

For each index $i$ such that $0\le i \le n$, there is a line bundle on $\overline{\mathcal{M}_{0,n}}(X,\beta)$ such that the fiber at a moduli point is the cotangent bundle of the curve at the $i^\mathrm{th}$ marked point. We denote by $\psi_i$ the first Chern class of this line bundle. In order to compute invariants involving ${\psi_1}^{a_1}\cdots {\psi_n}^{a_n}$, for some nonnegative integers $a_1,\ldots, a_n$, we write Psi(a_1,...,a_n)

Example

Let $G$ be the GKM graph of the Hirzebruch surface $\mathbb{P}(\mathcal{O}_{\mathbb{P}^1}(0) \oplus \mathcal{O}_{\mathbb{P}^1}(1))$, let $\beta$ the class of the fiber. The invariant

\[\int_{\overline{M}_{0,2}(G, \beta)}\mathrm{ev}_{1}^{*}([\mathrm{pt}])\cdot\psi_{1}^{0}\psi_{2} = -1,\]

can be computed as following.

julia> G = gkm_graph_of_toric(hirzebruch_surface(NormalToricVariety, 1));

julia> P = ev(1, point_class(G, 1)) * Psi(0,1);

julia> beta = curve_class(G, "1", "4"); # beta is a fiber of the map G -> P^1

julia> gromov_witten(G, beta, 2, P; show_bar=false)
-1